split up the fields.py into seperate files.
